Uploading illegal file will not trigger a change event
